requirements.txt
beautifulsoup4==4.7.1
gensim==3.6.0
irc==17.0
lxml==4.3.0
mysql-connector==2.1.6
networkx==2.2
nltk==3.4
pandas==0.23.4
numpy==1.15.4
plotly==3.5.0
regex==2018.1.10
requests==2.21.0
sklearn==0.0
textblob==0.15.2
ujson==1.35
urllib3==1.24.1
DateTime==4.3
